refactor: move some components to more adequate locations
