+++ src/components/sharedComponents/SwitchNetwork.tsx
@@ -5,7 +5,7 @@ import { Dropdown, Item as BaseItem } from '@bootnodedev/db-ui-toolkit'
 import * as chains from 'viem/chains'
 import { useSwitchChain } from 'wagmi'
 
-import { PrimaryButton } from '@/src/components/sharedComponents/Buttons'
+import { PrimaryButton } from '@/src/components/sharedComponents/ui/Buttons'
 import { useWeb3Status } from '@/src/hooks/useWeb3Status'
 
 type NetworkItem = {
