+++ src/components/sharedComponents/WalletStatusVerifier.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 
 import { extractChain } from 'viem'
 
-import { PrimaryButton } from '@/src/components/sharedComponents/Buttons'
+import { PrimaryButton } from '@/src/components/sharedComponents/ui/Buttons'
 import { useWeb3Status } from '@/src/hooks/useWeb3Status'
 import { chains, type ChainsIds } from '@/src/lib/networks.config'
 import { ConnectWalletButton } from '@/src/providers/Web3Provider'
