+++ src/routes/__root.tsx
@@ -5,10 +5,10 @@ import { Analytics } from '@vercel/analytics/react'
 import { ThemeProvider } from 'next-themes'
 import { Toaster } from 'react-hot-toast'
 
-import { Footer } from '@/src/components/sharedComponents/Footer'
-import { Header } from '@/src/components/sharedComponents/Header'
 import { TanStackReactQueryDevtools } from '@/src/components/sharedComponents/TanStackReactQueryDevtools'
 import { TanStackRouterDevtools } from '@/src/components/sharedComponents/TanStackRouterDevtools'
+import { Footer } from '@/src/components/sharedComponents/ui/Footer'
+import { Header } from '@/src/components/sharedComponents/ui/Header'
 import { TransactionNotificationProvider } from '@/src/lib/toast/TransactionNotificationProvider'
 import { Web3Provider } from '@/src/providers/Web3Provider'
 import Styles from '@/src/styles'
